Simplify Auth to single class
