+++ Backend/Common/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@inv2/common",
-  "version": "1.2.21",
+  "version": "1.2.22",
   "description": "A common module for v2",
   "main": "./build/index.js",
   "types": "./build/index.d.ts",
