+++ Backend/Common/src/middlewares/index.ts
@@ -1,6 +1,5 @@
 
-export * from './current-user';
+export * from './authentication';
 export * from './error-handler';
-export * from './require-auth';
 // export * from './validate-request';
 export * from './decorators/joi-validation.decorators';
